Keep new helper in separate file to prevent build errors
